+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.152.0] - 2026-08-17
+
+### Changes
+- feat(admin): support admins as regular volunteers, make user search index-backed ([#1203](https://github.com/everybody-eats-nz/volunteer-portal/pull/1203)) by @malinmalliyawadu
+
+
+## [0.151.9] - 2026-08-17
+
+### Changes
+- chore(deps): bump @types/jsdom from 28.0.3 to 30.0.0 in /web ([#1200](https://github.com/everybody-eats-nz/volunteer-portal/pull/1200)) by @dependabot[bot]
+
+
 ## [0.151.8] - 2026-08-11
 
 ### Changes